% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A_melanoma_v5}
\alias{A_melanoma_v5}
\title{Trichotomous immune cells and melanoma}
\description{
This is a trichotomous guide matrix (53 by 11) corresponding to the 10 immune cell types and melanoma
each row represents a gene. For the i-j entry, it is encoded as 1 when this gene i is NOT a signature gene to
celltype j; and it is encoded as 0.5 it is a secodnary siganture gene for the j cell type; and 0 if it is a primary
signature gene for j-th cell type. This encoding is counter-intuitive (if you think it should be the other way around)
is because it is easier to take this form for penalization.
}
