% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_structure_example}
\alias{cell_structure_example}
\title{Example of cell types matching table}
\description{
This is a dataframe that serves as reference when testing NITUMID's performance on datasets that with
different cell types specification. For example, if you have a melanoma dataset with known cell types proportions
for T cells, DC, Macrophages, Mature B cells (MB), NK cells and other cells together, then it does not match NITUMID's 11-celltype
setup. In this case, you could use this dataframe to specify the matching relationships between
your known cell types and NITUMID's cell types. This dataframe has two column, `origina_index` marks our 11 cell types
from 1 to 11, you are not supposed to modify that; you can modify `destin_index` accordingly as DC-1, CD4-2, CD8-2, Macrophages-3,
MB-4, NK-5, Monocytes-6, Plasma-6, MAST-6, Eosinophil-6, Melanoma-6. Basically, we merged CD4+ T cell and CD8+ T cell into one cell
type, and all our cell types after monocytes into the 6-th cell type.
\describe {
  \item{origina_index}{Indexed NITUMID's 11 cell types, from 1 to 11}
  \item{destin_index}{Indexed customized cell types }
}
}
