% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{NITUMID_simple}
\alias{NITUMID_simple}
\title{Simple version of NITUMID}
\usage{
NITUMID_simple(Y, if.bulk)
}
\arguments{
\item{Y}{Y is the gene expression matrix that has #{signature genes} rows and #{samples} columns, Y's rowname must be gene symbols}

\item{if.bulk}{Logical, if Y is bulk gene expression data. If Y is purified cell gene expression, please set it as FALSE.}
}
\value{
`NITUMID_simple` returns a list of two object, W matrix and H matrix
}
\description{
`NITUMID_simple` This is a simplified version of NITUMID, which uses most of the default settings (generic signature genes, etc.),
you can only input a gene expression data frame whose rownames is gene symbol, and tells it if this data is from pure cells or bulk tumor (melanoma)
}
\examples{
TBA

}
