% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{NMF_RNA_Tang}
\alias{NMF_RNA_Tang}
\title{NMF core function}
\usage{
NMF_RNA_Tang(Y, ini.W, ini.H, lam_W, beta_W, A, eta, tol = 10^-5,
  prterr = F, max.itr = 2000, always = T)
}
\arguments{
\item{Y}{Y is the gene expression matrix that has #{signature genes} rows and #{samples} columns}

\item{ini.W}{Initial W matrix}

\item{ini.H}{Initial H matrx}

\item{lam_W}{$\\lambda$ is the parameter imposing A's structure into resulting W, see our paper for more info. In
real calculation, upper level functions will automatically choose suitable $\\lambda$ no need to set by yourself}

\item{beta_W}{$\\beta$ is the parameter controlling sparsity ridge penalty, see our paper for more information. In
real calculation, upper level functions will automatically choose suitable $\\beta$ no need to set by yourself}

\item{A}{The trichotomous guide matrix A}

\item{eta}{$\\eta$ is the parameter for ensuring column sums of H are 1s, upper level function will choose suitable
$\\eta$ automatically, no need to spcify yourself}

\item{tol}{numeric scalar, the cutoff value for stop iterating, default $10^{-5}$}

\item{prterr}{logical value indicating if it should print each step's error, default FALSE}

\item{max.itr}{numerical scalar, maximum iteration number, default 2000}

\item{alaways}{logical, if it should output results whatsover when max.itr is reached, default TRUE}
}
\value{
'NMF_RNA_Tang' returns the resulted W and H, as well as track of errors along iteration
}
\description{
`NMF_RNA_Tang` is the core function for NITUMID's factorization, it takes the gene expression matrix Y,
guide matrix A, initial W and H, as well as other parameters as input, and optimizes til converged/ or
reached maximum steps
}
