% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{NMF_RNA_Tang_ind}
\alias{NMF_RNA_Tang_ind}
\title{NMF core function}
\usage{
NMF_RNA_Tang_ind(Y, lam_W, beta_W, A, eta, tol = 10^-5, prterr = F,
  max.itr = 2000, always = F, num_cell = 11, scale_Data,
  perm_type = T)
}
\arguments{
\item{Y}{Y is the gene expression matrix that has #{signature genes} rows and #{samples} columns}

\item{lam_W}{$\\lambda$ is the parameter imposing A's structure into resulting W, see our paper for more info. In
real calculation, upper level functions will automatically choose suitable $\\lambda$ no need to set by yourself}

\item{beta_W}{$\\beta$ is the parameter controlling sparsity ridge penalty, see our paper for more information. In
real calculation, upper level functions will automatically choose suitable $\\beta$ no need to set by yourself}

\item{A}{The trichotomous guide matrix A}

\item{eta}{$\\eta$ is the parameter for ensuring column sums of H are 1s, upper level function will choose suitable
$\\eta$ automatically, no need to spcify yourself}

\item{tol}{numeric scalar, the cutoff value for stop iterating, default $10^{-5}$}

\item{prterr}{logical value indicating if it should print each step's error, default FALSE}

\item{max.itr}{numerical scalar, maximum iteration number, default 2000}

\item{num_cell}{numeric scalar, number of component cell types, dfault 11. Noted that if you were to change component
cell types number, you should change your $A$ matrix accordingly.}

\item{scale_Data}{the parameter controlling which scaling method to use for $Y$: 0 indicates no scaling, 1 indicates
scaling $||Y||_{F}=\sqrt{nm}$, where $n,m$ are dimensions of Y; 2 indicates scaling such that $||Y^{scaled}||_{F}=||log(Y^{raw}+1)||_{F}$}

\item{perm_type}{logical variable, default TRUE. If TRUE, the function will do a permutation and measure consistency between results.}

\item{ini.W}{Initial W matrix}

\item{ini.H}{Initial H matrx}

\item{alaways}{logical, if it should output results whatsover when max.itr is reached, default TRUE}
}
\value{
'NMF_RNA_Tang' returns the resulted W and H, as well as track of errors along iteration
}
\description{
`NMF_RNA_Tang_ind` is a upper level wrapper function for `NMF_RNA_Tang`. On top of core NMF, it also initialtes $W$ and $H$, and scales/normalizes
according to the input arguments: scale_Data and reg_para. Besides, it also does random permutations to check results stability and helps choose the
best parameters. Although it has full sets of NITUMID's functions, the `NMF_RNA_Tang_ind` is still not supposed to be run independently, has hyper-function
will call it with different parameters choices.
}
\examples{
This function is not supposed to be run independently, it is supposed to be called by top level function, which will feed it with different pre-determined
parameters and choose thr best one

}
