% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{row_mean_v5}
\alias{row_mean_v5}
\title{Siganture genes' mean expression in training data}
\format{this is a numeric vector length 53}
\description{
This vector is the mean expression across all training data of 10 immune cell types, it is
to provide information on general relative gene expression level of all signature genes, which is used
in some normalizationg circumstances, see our paper for more details
}
