% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signature_marker_melanoma}
\alias{signature_marker_melanoma}
\title{Siganture genes' information for immune cells and melanoma}
\format{this is a dataframe with 53 rows and 4 variables:
\describe{
  \item{Cell}{to which cell type is this gene a primary signature}
  \item{Secondary_cells}{to which cell type(s) is this gene a secondary signature}
  \item{gene_symbol}{the gene symbol}
  \item{Supporting_Evidence}{links or publications that justify each gene as a siganture gene}
}}
\description{
A dataset contains all 53 signature genes' information for 10 immune cell types and
melanoma, including their gene symbols, corresponding primary and secondary cell types
and supporting evidence for each gene being signature. Notice that the sigantures are
subjective to changes between different tumor types. For a detailed introduction to primary and
secondary cell types, please refer to our paper.
}
