% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Signature_Match}
\alias{Signature_Match}
\title{Find signature genes' positions in gene expression data}
\usage{
Signature_Match(gene_expression, signature_genes,
  signature_genes_alias = NULL)
}
\arguments{
\item{gene_expression}{this should be a numeric matrix whose columns are samples and rows are genes,
its row names should be genes symbol}

\item{signature_genes}{this should be a vector of string/character speciying signature gene symbols
default value is our generic signature genes list}

\item{signature_genes_alias}{this optional argument allows you to give signature genes' alias, default is NULL.
this also should be a vector of string/character, same length as signature_genes, those who don't have an alias should
be set as NA.}
}
\value{
'Signature_Match' returns a list consists of two elements: 'matched_index' is a vector same length as signature genes
         its each entry specifying the corresponding signature genes' positions in 'gene_expression', for those genes don't present in
         data matrix, its corresponding entries are left with 1221; 'missing_row_index' is a vector indicating which genes are not found
         in 'gene_expression'
}
\description{
`Signature_Match` takes the input gene expression data and a vector of signature gene symbols, return signature
genes' positions in data matrix, as well as those not-matched signature genes.
}
\examples{
data is needed otherwise it won't run, the second example won't work because the alias vector has one non-character entry
(having NAs in signature_genes_alias is fine)
 Signature_Match(gene_expression_matrix,signature_genes=c("MIA","EGFR") )
 \dontrun{
 Signature_Match(gene_expression_matrix,signature_genes=c("MIA","EGFR"),signature_genes_alias=c(NA,2) )
 }
 @export
}
