% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{Update_H_Tang}
\alias{Update_H_Tang}
\title{Update of H matrix in iteration}
\usage{
Update_H_Tang(Y, W_itr, H_itr, eta)
}
\arguments{
\item{Y}{Y is the gene expression matrix that has #{signature genes} rows and #{samples} columns}

\item{W_itr}{W matrix from previous step of iteration}

\item{H_itr}{H matrix from previous step of iteration}

\item{eta}{$\\lambda$ is the parameter for ensuring column sums of H are 1s}

}
\value{
'Update_H_Tang' returns the udpated H matrix after iteration
}
\description{
`Update_H_Tang` conducts one simple iteration of H in the NMF process, however, it is designed as
a sub-function for NMF procedure, do not run it indenpendently
}
\examples{
This function is not supposed to be run independently
}
