% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{Update_W_Tang}
\alias{Update_W_Tang}
\title{Update of W matrix in iteration}
\usage{
Update_W_Tang(Y, W_itr, H_itr, beta_W, lam_W, A)
}
\arguments{
\item{Y}{Y is the gene expression matrix that has #{signature genes} rows and #{samples} columns}

\item{W_itr}{W matrix from previous step of iteration}

\item{H_itr}{H matrix from previous step of iteration}

\item{beta_W}{$\\beta$ is the parameter controlling sparsity ridge penalty, see our paper for more information}

\item{lam_W}{$\\lambda$ is the parameter imposing A's structure into resulting W, see our paper for more info}

\item{A}{The trichotomous guide matrix A}
}
\value{
'Update_W_Tang' returns the udpated W matrix after iteration
}
\description{
`Update_W_Tang` conducts one simple iteration of W in the NMF process, however, it is designed as
a sub-function for NMF procedure, do not run it indenpendently
}
\examples{
This function is not supposed to be run independently
}
